// src/types/direction.ts
export type Direction = 'top' | 'right' | 'bottom' | 'left';

export enum WallDirections {
  topLeft = 'topLeft',
  topMiddle = 'topMiddle',
  topRight = 'topRight',
  middleLeft = 'middleLeft',
  middleMiddle = 'middleMiddle',
  middleRight = 'middleRight',
  bottomLeft = 'bottomLeft',
  bottomMiddle = 'bottomMiddle',
  bottomRight = 'bottomRight',
  narrowTop = 'narrowTop',
  narrowRight = 'narrowRight',
  narrowBottom = 'narrowBottom',
  narrowLeft = 'narrowLeft',
  narrowHorizontal = 'narrowHorizontal',
  narrowVertical = 'narrowVertical',
  narrowMiddle = 'narrowMiddle',
}